Reimplement core-function using znc.in/server-time-iso
This is a re-write of the core function which makes use of of the ZNC
capabality znc.in/server-time-iso. It replaces the old method of parsing
appended timestamps which is cumbersome and error-prone. Plus the new
timestamps are guaranteed to be in UTC time so we can probably drop the
timestamp offset feature in the future.

For now this drops support for self-messages in queries.

Signed-off-by: Konrad Gräfe <konradgraefe@aol.com>