Re-implement by hooking into UI operations
This is faster than using the "writing-chat-msg" and "writing-im-msg"
signals as those do not allow us to modify the timestamp and therefore forced
us to cancel the received message and post a new one, which in turn
triggered all the signals again.

This also allows us to remove parted users with
purple_conv_chat_remove_users() instead of tinkering with the GtkHTML
ourselves.

Signed-off-by: Konrad Gräfe <konradgraefe@aol.com>