Fix setting capabilities in upcoming Pidgin 2.14.10
Pidgin 2.14.9 introduced message rate limiting for IRC. With that
messages sent with send_raw() will be queued up before they are sent
eventually. This is too late for capability negotiation.

Instead we modify the message being sent with priority so our
capabilities are sent immediately.

Note that this does not work with Pidgin 2.14.9 due to
https://issues.imfreedom.org/issue/PIDGIN-17626

Signed-off-by: Konrad Gräfe <kgraefe@paktolos.net>